src/services/ai-floorplan/requirements.txt
numpy>=2.0.0,<3.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.12,<1.0.0
opencv-python-headless>=4.10.0,<5.0.0
pydantic>=2.0.0,<3.0.0
pillow>=10.0.0,<12.0.0
inference-sdk>=0.9.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
